refactor: separate web app routing into router.ts
- Moved `doGet` from `main.ts` to `router.ts`.
- Implemented `doPost` in `router.ts` for future webhook expansions.
- Moved `importPastActivitiesFromWeb` from `manual_import.ts` to `router.ts`.
- Added `ContentService` mock to `vitest.setup.ts`.
- Migrated relevant tests to `tests/router.spec.ts`.
- Cleaned up exports and unused functions in `main.ts` and `manual_import.ts`.
