+++ manual_import.ts
@@ -1,46 +1,4 @@
 // ==========================================
-// 【Webアプリ用】画面から受け取った日付でインポートを実行
-// ==========================================
-function importPastActivitiesFromWeb(startStr: string, endStr: string): string {
-    // Validate input format YYYY-MM-DD
-    const dateRegex = /^\d{4}-\d{2}-\d{2}$/;
-    if (!startStr || !endStr || !dateRegex.test(startStr) || !dateRegex.test(endStr)) {
-        const msg = 'エラー: 日付の形式が正しくありません (YYYY-MM-DD)。';
-        Logger.log(msg);
-        return msg;
-    }
-
-    // Helper to validate date components to prevent rollover (e.g., 2024-02-31 -> 2024-03-02)
-    function isValidDateComponents(dateStr: string, dateObj: Date): boolean {
-        if (isNaN(dateObj.getTime())) return false;
-        const [y, m, d] = dateStr.split('-');
-        return dateObj.getFullYear() === parseInt(y, 10) &&
-            dateObj.getMonth() + 1 === parseInt(m, 10) &&
-            dateObj.getDate() === parseInt(d, 10);
-    }
-
-    // 画面からの文字列(YYYY-MM-DD)をDateオブジェクトに変換
-    const startDate = new Date(`${startStr}T00:00:00`);
-    const endDate = new Date(`${endStr}T23:59:59`);
-
-    // Check if dates are valid
-    if (!isValidDateComponents(startStr, startDate) || !isValidDateComponents(endStr, endDate)) {
-        const msg = 'エラー: 無効な日付が指定されました。';
-        Logger.log(msg);
-        return msg;
-    }
-
-    if (startDate > endDate) {
-        const msg = 'エラー: 開始日は終了日より前の日付を指定してください。';
-        Logger.log(msg);
-        return msg;
-    }
-
-    return importPastActivities(startDate, endDate);
-}
-
-
-// ==========================================
 // 指定した期間の過去データを取り込む (画面からの実行用)
 // ==========================================
 function importPastActivities(startDate?: Date, endDate?: Date, perPage: number = 200): string {
@@ -115,6 +73,5 @@ function importPastActivities(startDate?: Date, endDate?: Date, perPage: number 
 if (typeof module !== 'undefined' && module.exports) {
     module.exports = {
         importPastActivities,
-        importPastActivitiesFromWeb,
     };
 }
