+++ tests/manual_import.spec.ts
@@ -1,47 +1,7 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
-import { importPastActivities, importPastActivitiesFromWeb } from '../manual_import';
+import { importPastActivities } from '../manual_import';
 
 describe('manual_import', () => {
-
-    describe('importPastActivitiesFromWeb', () => {
-        beforeEach(() => {
-            vi.resetAllMocks();
-            vi.stubGlobal('Logger', { log: vi.fn() });
-            vi.stubGlobal('getStravaActivities', vi.fn());
-            vi.stubGlobal('getTargetCalendar', vi.fn());
-        });
-
-    afterEach(() => {
-        vi.unstubAllGlobals();
-    });
-
-        it('should correctly parse dates and pass them to importPastActivities', () => {
-            global.getStravaActivities.mockReturnValue([]);
-
-            const startStr = '2024-03-01';
-            const endStr = '2024-03-31';
-
-            const result = importPastActivitiesFromWeb(startStr, endStr);
-
-            expect(result).toBe('該当する期間のアクティビティはありませんでした。');
-
-            // Check that getStravaActivities was called with correctly parsed dates
-            expect(global.getStravaActivities).toHaveBeenCalledTimes(1);
-            const callArgs = global.getStravaActivities.mock.calls[0];
-
-            expect(callArgs[0]).toBeInstanceOf(Date);
-            // using getTime() to compare properly, accounting for local timezone issues where Date string output varies
-            const expectedStart = new Date('2024-03-01T00:00:00');
-            const expectedEnd = new Date('2024-03-31T23:59:59');
-
-            expect(callArgs[0].getTime()).toBe(expectedStart.getTime());
-
-            expect(callArgs[1]).toBeInstanceOf(Date);
-            expect(callArgs[1].getTime()).toBe(expectedEnd.getTime());
-
-            expect(callArgs[2]).toBe(200); // default perPage
-        });
-    });
 
     beforeEach(() => {
         vi.resetAllMocks();
@@ -138,56 +98,3 @@ describe('manual_import', () => {
     });
 });
 
-describe('importPastActivitiesFromWeb', () => {
-    beforeEach(() => {
-        vi.resetAllMocks();
-        vi.stubGlobal('getStravaActivities', vi.fn().mockReturnValue([]));
-        vi.stubGlobal('getTargetCalendar', vi.fn().mockReturnValue(null));
-        vi.stubGlobal('Logger', { log: vi.fn() });
-    });
-
-    it('should reject invalid date formats', () => {
-        const result1 = importPastActivitiesFromWeb('2024/01/01', '2024-01-31');
-        const result2 = importPastActivitiesFromWeb('2024-01-01', '2024/01/31');
-        const result3 = importPastActivitiesFromWeb('abcdef', '123456');
-        const result4 = importPastActivitiesFromWeb('', '2024-01-31');
-
-        const expectedError = 'エラー: 日付の形式が正しくありません (YYYY-MM-DD)。';
-        expect(result1).toBe(expectedError);
-        expect(result2).toBe(expectedError);
-        expect(result3).toBe(expectedError);
-        expect(result4).toBe(expectedError);
-        expect(global.Logger.log).toHaveBeenCalledWith(expectedError);
-    });
-
-    it('should reject invalid dates', () => {
-        const result = importPastActivitiesFromWeb('2024-13-45', '2024-01-31');
-        expect(result).toBe('エラー: 無効な日付が指定されました。');
-        expect(global.Logger.log).toHaveBeenCalledWith('エラー: 無効な日付が指定されました。');
-    });
-
-    it('should reject invalid dates with rollover', () => {
-        const result1 = importPastActivitiesFromWeb('2024-02-31', '2024-03-31');
-        expect(result1).toBe('エラー: 無効な日付が指定されました。');
-
-        const result2 = importPastActivitiesFromWeb('2024-11-31', '2024-12-01');
-        expect(result2).toBe('エラー: 無効な日付が指定されました。');
-    });
-
-    it('should reject date ranges where start is after end', () => {
-        const result = importPastActivitiesFromWeb('2024-01-31', '2024-01-01');
-        expect(result).toBe('エラー: 開始日は終了日より前の日付を指定してください。');
-        expect(global.Logger.log).toHaveBeenCalledWith('エラー: 開始日は終了日より前の日付を指定してください。');
-    });
-
-    it('should process valid dates', () => {
-        const startStr = '2024-01-01';
-        const endStr = '2024-01-31';
-
-        global.getStravaActivities.mockReturnValue([]);
-
-        const result = importPastActivitiesFromWeb(startStr, endStr);
-        expect(result).toBe('該当する期間のアクティビティはありませんでした。');
-        expect(global.getStravaActivities).toHaveBeenCalled();
-    });
-});
